+++ crates/dbflux_core/src/driver/mod.rs
@@ -8,8 +8,7 @@ pub use capabilities::{
     SslCertFields, SslModeOption, SyntaxInfo, TransactionCapabilities, WhereOperator,
 };
 pub use form::{
-    CLOUDWATCH_FORM, DYNAMODB_FORM, DriverFormDef, FormFieldDef, FormFieldKind, FormSection,
-    FormTab, FormValues, INFLUXDB_FORM, MONGODB_FORM, MYSQL_FORM, POSTGRES_FORM, REDIS_FORM,
-    RefreshTrigger, SQLITE_FORM, SQLSERVER_FORM, SelectOption, field_file_path, field_password,
-    field_use_uri, ssh_tab,
+    DriverFormDef, FormFieldDef, FormFieldKind, FormSection, FormTab, FormValues, RefreshTrigger,
+    SelectOption, field, field_file_path, field_password, field_required, field_use_uri, ssh_tab,
+    when_checked, when_unchecked, with_default, with_help,
 };
