+++ crates/dbflux_core/src/lib.rs
@@ -91,14 +91,13 @@ pub use data::{
 };
 
 pub use driver::{
-    CLOUDWATCH_FORM, DYNAMODB_FORM, DatabaseCategory, DdlCapabilities, DeploymentClass,
-    DriverCapabilities, DriverFormDef, DriverLimits, DriverMetadata, DriverMetadataBuilder,
-    ExecutionClassification, FormFieldDef, FormFieldKind, FormSection, FormTab, FormValues,
-    INFLUXDB_FORM, Icon, IsolationLevel, MONGODB_FORM, MYSQL_FORM, MutationCapabilities,
-    OperationClassifier, POSTGRES_FORM, PaginationStyle, QueryCapabilities, QueryLanguage,
-    REDIS_FORM, RefreshTrigger, SQLITE_FORM, SQLSERVER_FORM, SelectOption, SslCertFields,
-    SslModeOption, SyntaxInfo, TransactionCapabilities, WhereOperator, field_file_path,
-    field_password, field_use_uri, ssh_tab,
+    DatabaseCategory, DdlCapabilities, DeploymentClass, DriverCapabilities, DriverFormDef,
+    DriverLimits, DriverMetadata, DriverMetadataBuilder, ExecutionClassification, FormFieldDef,
+    FormFieldKind, FormSection, FormTab, FormValues, Icon, IsolationLevel, MutationCapabilities,
+    OperationClassifier, PaginationStyle, QueryCapabilities, QueryLanguage, RefreshTrigger,
+    SelectOption, SslCertFields, SslModeOption, SyntaxInfo, TransactionCapabilities, WhereOperator,
+    field, field_file_path, field_password, field_required, field_use_uri, ssh_tab, when_checked,
+    when_unchecked, with_default, with_help,
 };
 
 pub use facade::{DangerousQuerySuppressions, SessionFacade};
