+++ crates/dbflux_driver_cloudwatch/src/lib.rs
@@ -12,4 +12,4 @@
 pub mod driver;
 pub mod metric_catalog;
 
-pub use driver::{CLOUDWATCH_METADATA, CloudWatchDriver};
+pub use driver::{CLOUDWATCH_FORM, CLOUDWATCH_METADATA, CloudWatchDriver};
