+++ crates/dbflux_driver_mongodb/src/lib.rs
@@ -14,7 +14,7 @@ pub mod language_service;
 pub mod query_generator;
 pub mod query_parser;
 
-pub use driver::{MONGODB_METADATA, MongoDriver};
+pub use driver::{MONGODB_FORM, MONGODB_METADATA, MongoDriver};
 pub use language_service::MongoLanguageService;
 pub use query_generator::MongoShellGenerator;
 pub use query_parser::{MongoParseError, validate_query, validate_query_positional};
