+++ crates/dbflux_driver_mssql/src/lib.rs
@@ -12,5 +12,5 @@
 pub mod driver;
 pub mod language_service;
 
-pub use driver::{METADATA, MssqlDriver};
+pub use driver::{METADATA, MssqlDriver, SQLSERVER_FORM};
 pub use language_service::TSqlLanguageService;
