+++ crates/dbflux_driver_redis/src/lib.rs
@@ -14,5 +14,5 @@ pub mod driver;
 pub mod language_service;
 
 pub use command_generator::RedisCommandGenerator;
-pub use driver::{REDIS_METADATA, RedisDriver};
+pub use driver::{REDIS_FORM, REDIS_METADATA, RedisDriver};
 pub use language_service::RedisLanguageService;
